+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 0.1.5
     - Dispatch only published posts
+    - Added ID getter and setter for Posts
 
 0.1.4
     - Application handles only outside of admin panel